processor-python/requirements.txt
pytesseract==0.3.10
Pillow==9.5.0
transformers==4.30.2
torch==2.7.0
PyPDF2==3.0.1
pdf2image==1.16.3
reportlab==4.0.4
tqdm>=4.65.0
